+++ src/main/java/io/supertokens/webserver/api/emailpassword/SignUpAPI.java
@@ -68,7 +68,11 @@ public class SignUpAPI extends WebserverAPI {
 
             JsonObject result = new JsonObject();
             result.addProperty("status", "OK");
-            result.add("user", new JsonParser().parse(new Gson().toJson(user)).getAsJsonObject());
+            JsonObject userJson = new JsonParser().parse(new Gson().toJson(user)).getAsJsonObject();
+            if (super.getVersionFromRequest(req).equals("2.4")) {
+                userJson.remove("timeJoined");
+            }
+            result.add("user", userJson);
             super.sendJsonResponse(200, result, resp);
 
         } catch (DuplicateEmailException e) {